Fixed bug as a result of changes in libsnert 1.72.5 concerning the separation of IO errors from not-found answers in smdb API (made for BarricadeMX use of socketmaps).
The call-ahead-db option uses the smdb API. When disabled, the
call to smdbAccessDomain used to returned SMDB_ACCESS_NOT_FOUND,
but now returns SMDB_ACCESS_TEMPFAIL with a "TEMPFAIL" string on
error.

call-ahead-db values are similar to mailertable, thus smdbAccess
routinues should always return SMDB_ACCESS_UNKNOWN and a string;
anything else is invalid for call-ahead-db and can be ignored.